지역함수 파일 분리
