+++ src/constants/location.constant.ts
@@ -1,40 +1,102 @@
-import { Price } from '../@types/list/listType';
-
-export type CheckboxOptionType = { value: string; text: string; description?: string };
-
-export const PERSONALITY_LIST: CheckboxOptionType[] = [
-  { value: 'intellectual', text: '지적인' },
-  { value: 'calm', text: '차분한' },
-  { value: 'humorous', text: '유머있는' },
-  { value: 'optimistic', text: '낙천적인' },
-  { value: 'passionate', text: '열정적인' },
-  { value: 'emotional', text: '감성적인' },
-  { value: 'introverted', text: '내향적인' },
-  { value: 'extroverted', text: '외향적인' },
-  { value: 'kind', text: '상냥한' },
-  { value: 'unique', text: '개성있는' },
-];
-
-export const AVAILABLE_LANGUAGE_LIST: CheckboxOptionType[] = [
-  { value: 'KR', text: '한국어' },
-  { value: 'JP', text: '일본어' },
-  { value: 'CH', text: '중국어' },
-];
-
-export const CLASSLEVEL_LIST: CheckboxOptionType[] = [
-  { value: 'Beginner', text: '초급', description: '기본적인 인사와 간단한 일상 대화를 배우며 기초적인 문법과 어휘를 익히는 단계' },
-  { value: 'Intermediate', text: '중급', description: '좀 더 복잡한 주제에 대해 의사소통하고 일상 대화 뿐만 아니라 간단한 토론과 의견 표현을 연습하는 단계.' },
-  { value: 'Advanced', text: '고급', description: '심도 있는 토론과 주제에 대한 깊은 이해를 바탕으로 논리적인 의견을 표현하고 어려운 어휘와 문법을 다루는 단계' },
-];
-
-export const ENROLLMENT_STATUS: CheckboxOptionType[] = [
-  { value: 'enrolled', text: '재학중' },
-  { value: 'onLeave', text: '휴학중' },
-  { value: 'graduated', text: '졸업' },
-  { value: 'withdrawn', text: '중퇴' },
-];
-
 export const AREA0 = ['전체', '서울', '인천', '대전', '광주', '대구', '울산', '부산', '경기', '강원', '충북', '충남', '전북', '전남', '경북', '경남', '제주'];
+export const AREA1 = [
+  '강남구',
+  '강동구',
+  '강북구',
+  '강서구',
+  '관악구',
+  '광진구',
+  '구로구',
+  '금천구',
+  '노원구',
+  '도봉구',
+  '동대문구',
+  '동작구',
+  '마포구',
+  '서대문구',
+  '서초구',
+  '성동구',
+  '성북구',
+  '송파구',
+  '양천구',
+  '영등포구',
+  '용산구',
+  '은평구',
+  '종로구',
+  '중구',
+  '중랑구',
+];
+export const AREA2 = ['계양구', '남구', '남동구', '동구', '부평구', '서구', '연수구', '중구', '강화군', '옹진군'];
+export const AREA3 = ['대덕구', '동구', '서구', '유성구', '중구'];
+export const AREA4 = ['광산구', '남구', '동구', '북구', '서구'];
+export const AREA5 = ['남구', '달서구', '동구', '북구', '서구', '수성구', '중구', '달성군'];
+export const AREA6 = ['남구', '동구', '북구', '중구', '울주군'];
+export const AREA7 = ['강서구', '금정구', '남구', '동구', '동래구', '부산진구', '북구', '사상구', '사하구', '서구', '수영구', '연제구', '영도구', '중구', '해운대구', '기장군'];
+export const AREA8 = [
+  '고양시',
+  '과천시',
+  '광명시',
+  '광주시',
+  '구리시',
+  '군포시',
+  '김포시',
+  '남양주시',
+  '동두천시',
+  '부천시',
+  '성남시',
+  '수원시',
+  '시흥시',
+  '안산시',
+  '안성시',
+  '안양시',
+  '양주시',
+  '오산시',
+  '용인시',
+  '의왕시',
+  '의정부시',
+  '이천시',
+  '파주시',
+  '평택시',
+  '포천시',
+  '하남시',
+  '화성시',
+  '가평군',
+  '양평군',
+  '여주군',
+  '연천군',
+];
+export const AREA9 = ['강릉시', '동해시', '삼척시', '속초시', '원주시', '춘천시', '태백시', '고성군', '양구군', '양양군', '영월군', '인제군', '정선군', '철원군', '평창군', '홍천군', '화천군', '횡성군'];
+export const AREA10 = ['제천시', '청주시', '충주시', '괴산군', '단양군', '보은군', '영동군', '옥천군', '음성군', '증평군', '진천군', '청원군'];
+export const AREA11 = ['계룡시', '공주시', '논산시', '보령시', '서산시', '아산시', '천안시', '금산군', '당진군', '부여군', '서천군', '연기군', '예산군', '청양군', '태안군', '홍성군'];
+export const AREA12 = ['군산시', '김제시', '남원시', '익산시', '전주시', '정읍시', '고창군', '무주군', '부안군', '순창군', '완주군', '임실군', '장수군', '진안군'];
+export const AREA13 = ['광양시', '나주시', '목포시', '순천시', '여수시', '강진군', '고흥군', '곡성군', '구례군', '담양군', '무안군', '보성군', '신안군', '영광군', '영암군', '완도군', '장성군', '장흥군', '진도군', '함평군', '해남군', '화순군'];
+export const AREA14 = [
+  '경산시',
+  '경주시',
+  '구미시',
+  '김천시',
+  '문경시',
+  '상주시',
+  '안동시',
+  '영주시',
+  '영천시',
+  '포항시',
+  '고령군',
+  '군위군',
+  '봉화군',
+  '성주군',
+  '영덕군',
+  '영양군',
+  '예천군',
+  '울릉군',
+  '울진군',
+  '의성군',
+  '청도군',
+  '청송군',
+  '칠곡군',
+];
+export const AREA15 = ['거제시', '김해시', '마산시', '밀양시', '사천시', '양산시', '진주시', '진해시', '창원시', '통영시', '거창군', '고성군', '남해군', '산청군', '의령군', '창녕군', '하동군', '함안군', '함양군', '합천군'];
+export const AREA16 = ['서귀포시', '제주시', '남제주군', '북제주군'];
 
 export const 서울 = [
   '전체',
@@ -138,17 +200,3 @@ export const 경북 = [
 
 export const 경남 = ['전체', '거제시', '김해시', '마산시', '밀양시', '사천시', '양산시', '진주시', '진해시', '창원시', '통영시', '거창군', '고성군', '남해군', '산청군', '의령군', '창녕군', '하동군', '함안군', '함양군', '합천군'];
 export const 제주 = ['전체', '서귀포시', '제주시', '남제주군', '북제주군'];
-
-export const gender: string[] = ['전체', '여성', '남성'];
-export const level: string[] = ['전체', '초급', '중급', '고급'];
-export const age: string[] = ['전체', '10대', '20대', '30대', '40대', '50대'];
-export const speakingLanguage: string[] = ['전체', '한국어', '일본어', '중국어'];
-
-export const price: Price[] = [
-  { optionPrice: '전체', min: 0, max: 100000 },
-  { optionPrice: '5,000 ~ 10,000', min: 5000, max: 10000 },
-  { optionPrice: '10,000 ~ 20,000', min: 10000, max: 20000 },
-  { optionPrice: '20,000 ~ 30,000', min: 20000, max: 30000 },
-  { optionPrice: '30,000 ~ 40,000', min: 30000, max: 40000 },
-  { optionPrice: '50,000 ~', min: 50000, max: 100000 },
-];
